Refactor CallView environment initialization
