+++ GliaWidgets/Sources/View/Call/CallView.Environment.swift
@@ -1,0 +1,29 @@
+import Foundation
+
+extension CallView {
+    struct Environment {
+        var data: FoundationBased.Data
+        var uuid: () -> UUID
+        var gcd: GCD
+        var imageViewCache: ImageView.Cache
+        var timerProviding: FoundationBased.Timer.Providing
+        var uiApplication: UIKitBased.UIApplication
+        var uiScreen: UIKitBased.UIScreen
+        var uiDevice: UIKitBased.UIDevice
+    }
+}
+
+extension CallView.Environment {
+    static func create(with environment: ViewFactory.Environment) -> Self {
+        .init(
+            data: environment.data,
+            uuid: environment.uuid,
+            gcd: environment.gcd,
+            imageViewCache: environment.imageViewCache,
+            timerProviding: environment.timerProviding,
+            uiApplication: environment.uiApplication,
+            uiScreen: environment.uiScreen,
+            uiDevice: environment.uiDevice
+        )
+    }
+}
