+++ GliaWidgets/Sources/View/Call/CallView.swift
@@ -523,16 +523,3 @@ extension CallView {
         )
     }
 }
-
-extension CallView {
-    struct Environment {
-        var data: FoundationBased.Data
-        var uuid: () -> UUID
-        var gcd: GCD
-        var imageViewCache: ImageView.Cache
-        var timerProviding: FoundationBased.Timer.Providing
-        var uiApplication: UIKitBased.UIApplication
-        var uiScreen: UIKitBased.UIScreen
-        var uiDevice: UIKitBased.UIDevice
-    }
-}
